fastapi_microservice/requirements.txt
aioredis==2.0.1
sqlalchemy==2.0.32
fastapi==0.112.2
uvicorn==0.30.6
psycopg2==2.9.9
profi_log==0.1.0
fastapi-cache2==0.2.2
redis==5.0.8